+++ android/src/main/java/com/jiguang/jverify/JverifyPlugin.java
@@ -40,7 +40,6 @@ import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
 
-
 /**
  * JverifyPlugin
  */