+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 3.0.0
++ 新增：支持空安全；
++ 优化：升级 AGP 及 Gradle 版本；
++ 优化：compileSdkVersion/targetSdkVersion 28 -> 30；
++ 优化：迁移至 AndroidX；
++ 优化：移除 jcenter；
++ 优化：移除废弃方法。
+### 破坏性改动
++ `addClikWidgetEventListener` -> `addClickWidgetEventListener`
 ## 2.0.7
 + 修复：修复android gradle 升级4.0 以上版本编译问题
 ## 2.0.5